+++ derive/tests/ui/enum_mock.rs
@@ -1,0 +1,9 @@
+use mimicry_derive::CallReal;
+
+#[derive(CallReal)]
+enum MyMock {
+    Some(u32),
+    None,
+}
+
+fn main() {}