+++ derive/tests/ui/mock_with_multiple_implicit_switches.rs
@@ -1,0 +1,12 @@
+use mimicry_derive::CallReal;
+
+/// Dummy struct to trick `CallReal` derive logic.
+struct RealCallSwitch;
+
+#[derive(CallReal)]
+struct MyMock {
+    switch: RealCallSwitch,
+    another_switch: RealCallSwitch,
+}
+
+fn main() {}