+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 - Support passing values to `Answers` via a message channel. This allows
   controlling when the answers are consumed and specifying answers after
   the mock is set.
+- Support mocking async functions / methods.
+
+### Changed
+
+- Change `call_real()` / `call_real_once()` interface. Now, these methods return
+  a guard that can then be used on its own or using `scope()` / `async_scope()` wrappers.
+- Bump minimum supported Rust version from 1.57 to 1.59.
 
 ## 0.1.0 - 2022-07-04